requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
opentelemetry-exporter-jaeger
opentelemetry-exporter-digma
requests
